requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
pymongo==4.11.2
python-dotenv==1.0.1
motor==3.7.0